third_party/WebKit/LayoutTests/editing/deleting/smart-delete-004.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<p>Please tun this test with DRT.</p>
<p>To run it manually, double click on 'bar' and hit delete.  You should see 'foo baz'.</p>
<div id="includeSpace" contenteditable="true">foo b<span id="target">a</span>r baz</div>
<div id="notIncludeSpace" contenteditable="true">foo b<span id="target">a</span>r baz</div>
<script>
description('Test for smart delete forward on double click selected word.');
var sample;
function testIt(sampleId, expectedString, expectedOffset) {
    if (!window.testRunner)
      return;
    sample = document.getElementById(sampleId);
    var target = sample.querySelector('#target');
    internals.settings.setSelectTrailingWhitespaceEnabled(sampleId == 'includeSpace');
    eventSender.mouseMoveTo(target.offsetLeft, target.offsetTop);
    eventSender.mouseDown();
    eventSender.mouseUp();
    eventSender.mouseDown();
    eventSender.mouseUp();
    debug(sampleId);
    testRunner.execCommand('DeleteForward');
    shouldBeEqualToString('sample.innerText', expectedString);
    shouldBe('getSelection().rangeCount', '1');
    shouldBe('getSelection().focusNode',  'sample.firstChild');
    shouldBe('getSelection().focusOffset', expectedOffset.toString());
}
testIt('includeSpace', 'foo baz', 4);
testIt('notIncludeSpace', 'foo\u00A0baz', 3);
</script>
