third_party/WebKit/LayoutTests/editing/deleting/smart-delete-004-expected.txt

Test for smart delete forward on double click selected word.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


includeSpace
PASS sample.innerText is "foo baz"
PASS getSelection().rangeCount is 1
PASS getSelection().focusNode is sample.firstChild
PASS getSelection().focusOffset is 4
notIncludeSpace
PASS sample.innerText is "foo baz"
PASS getSelection().rangeCount is 1
PASS getSelection().focusNode is sample.firstChild
PASS getSelection().focusOffset is 3
PASS successfullyParsed is true

TEST COMPLETE
Please tun this test with DRT.

To run it manually, double click on 'bar' and hit delete. You should see 'foo baz'.

foo baz
foo baz
